+++ breeze-core/src/test/kotlin/com/windea/breezeframework/core/extensions/CollectionExtensionsKtTest.kt
@@ -6,8 +6,8 @@ class CollectionExtensionsKtTest {
 	@Test
 	fun repeatExtensionTest() {
 		println(listOf(1, 2, 3).repeat(3))
-		println(listOf(1, 2, 3).flatRepeat(3))
-		println(listOf(1, 2, 3).repeatChunked(3))
+		println(listOf(1, 2, 3).repeatOrdinal(3))
+		//println(listOf(1, 2, 3).repeatChunked(3))
 	}
 
 	@Test