+++ CHANGELOG.md
@@ -284,6 +284,7 @@
 * [X] breeze-core 重命名一些方法名，将一些方法改为顺序，优化一些代码逻辑。
 * [X] breeze-core 为`SystemExtensions`添加`executeCommand`扩展。
 * [X] breeze-core 添加一些字符串集合的特殊处理扩展。
+* [X] breeze-core 重构字符串和集合类型的`repeat`相关扩展。
 * [ ] breeze-core 让集合的`deepQuery`方法当特定的占位符不匹配类型时，回调为普通字符串。 
 * [ ] breeze-unstable 新的集合类型：`Tree`。
 * [ ] breeze-unstable 新的集合类型：`Table`。